requirements.txt
numpy==1.24.2
python_igraph==0.10.4
setuptools==65.5.0
